this is push version 1.0

+++ ex01/RPN.hpp
@@ -9,7 +9,7 @@ class RPN
 private:
     int operators;
     int tokens;
-    std::stack<int> numbers;
+    std::stack<double> numbers;
 public:
     RPN();
     ~RPN();
